+++ instrumentation/github.com/jackc/pgx/splunkpgx/sql.go
@@ -43,7 +43,7 @@ import (
 	"net/url"
 	"strings"
 
-	"github.com/jackc/pgx"
+	pgx "github.com/jackc/pgx/v4"
 	// Make sure to import this so the instrumented driver is registered.
 	_ "github.com/jackc/pgx/v4/stdlib"
 
@@ -61,7 +61,11 @@ func init() { // nolint: gochecknoinits
 // Postgres database using the github.com/jackc/pgx client package.
 func DSNParser(dataSourceName string) (splunksql.ConnectionConfig, error) {
 	var connCfg splunksql.ConnectionConfig
-	c, err := pgx.ParseConnectionString(dataSourceName)
+	// ParseConfig defaults:
+	//   host: OS specific unix path, `localhost` otherwise
+	//   port: 5432
+	//   user: OS user name
+	c, err := pgx.ParseConfig(dataSourceName)
 	if err != nil {
 		return connCfg, err
 	}