+++ CHANGELOG.md
@@ -31,6 +31,13 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
   environment variable to `jaeger-thrift-splunk`, or by directly passing the
   user configured exporter with a `WithTraceExporter` option. (#300)
 
+### Fixed
+
+- Consistently import `github.com/jackc/pgx/v4`, instead of
+  `github.com/jackc/pgx`, in the
+  `github.com/signalfx/splunk-otel-go/instrumentation/github.com/jackc/pgx/splunkpgx`
+  instrumentation. (#478)
+
 ## [0.7.0] - 2022-01-13
 
 ### Added